refactor(daemon): extract memory candidate queries
